[Bug] Match weight dtype to input dtype in AR+RMS(+QUANT) replacements
The AITER fused all-reduce kernels (fused_ar_rms and
fused_ar_rms_per_group_quant) take a single opaque pointer for the
norm weight and require its element dtype to equal the activation
dtype. Models such as Qwen3-Next register the RMSNorm weight as
float32 even when the activations are bf16, so the existing
replacements -- both the AR+RMS-only patterns and the three new
AR+RMS+per-group-FP8-quant patterns from this PR -- handed the kernel
a mismatched (fp32 weight, bf16 input) pair that silently corrupted
the normed activations. The downstream effect is poor GSM8K / MTP
accuracy on Qwen3-Next TP2/TP4; TP1 was unaffected only because the
AR+RMS fusion never fires in that case.

Tres Popp identified the root cause and fixed the two existing
patterns in `tpopp/aiter-ar-rmsnorm-weight-dtype`. This commit:

* Picks up Tres's two-line `weight.to(input.dtype)` fix in
  `AiterAllreduceFusedRMSNormPattern` and
  `AiterAllreduceFusedAddRMSNormPattern`, and
* Extends the same cast defensively to the three new AR+RMS+per-group-
  FP8-quant patterns introduced earlier in this PR
  (`AiterAllreduceFusedRMSNormGroupQuantFP8Pattern`,
  `AiterAllreduceFusedAddRMSNormGroupQuantFP8Pattern`, and
  `AiterAllreduceFusedAddRMSNormGroupQuantWithIndexerPattern`), all
  of which forward weight straight into the same
  `aiter_ar.fused_ar_rms_per_group_quant` kernel and thus inherit the
  same dtype constraint.

When weight already matches input.dtype the `.to()` is a no-op (and
constant-folds at compile time for graph-input parameters), so other
models pay nothing.

Diagnosis credit: Tres Popp, Nico Holmberg (reproduction +
confirmation on Qwen3-Next-80B-A3B-Instruct-FP8 TP2).

+++ vllm/compilation/passes/fusion/allreduce_rms_fusion.py
@@ -951,7 +951,7 @@ class AiterAllreduceFusedRMSNormPattern(BasePattern, VllmPatternReplacement):
             allreduce = self.FUSED_AR_RMSNORM_OP(
                 input_=input,
                 residual=residual,
-                weight=weight,
+                weight=weight.to(input.dtype),
                 epsilon=self.epsilon,
             )
             return allreduce[0], allreduce[1]
@@ -997,7 +997,7 @@ class AiterAllreduceFusedAddRMSNormPattern(BasePattern, VllmPatternReplacement):
             allreduce = self.FUSED_AR_RMSNORM_OP(
                 input_=input,
                 residual=residual,
-                weight=weight,
+                weight=weight.to(input.dtype),
                 epsilon=self.epsilon,
             )
             return allreduce[0], allreduce[1]
@@ -1074,7 +1074,7 @@ class AiterAllreduceFusedRMSNormGroupQuantFP8Pattern(
             result = self.FUSED_AR_RMS_QUANT_OP(
                 input_=input,
                 residual=residual,
-                weight=weight,
+                weight=weight.to(input.dtype),
                 epsilon=self.epsilon,
                 group_size=self.group_size,
             )
@@ -1154,7 +1154,7 @@ class AiterAllreduceFusedAddRMSNormGroupQuantFP8Pattern(
             result = self.FUSED_AR_RMS_QUANT_OP(
                 input_=input,
                 residual=residual,
-                weight=weight,
+                weight=weight.to(input.dtype),
                 epsilon=self.epsilon,
                 group_size=self.group_size,
             )
@@ -1271,7 +1271,7 @@ class AiterAllreduceFusedAddRMSNormGroupQuantWithIndexerPattern(
             fused = self.FUSED_AR_RMS_QUANT_BF16_OP(
                 input_=input_,
                 residual=residual,
-                weight=norm_weight,
+                weight=norm_weight.to(input_.dtype),
                 epsilon=eps,
                 group_size=gs,
             )
